+++ shotmanager/properties/shot.py
@@ -61,7 +61,7 @@ def list_shot_types(self, context):
     icon_stb = config.icons_col["ShotManager_CamGPStb_32"]
     res = (
         ("PREVIZ", "Camera Shot", "Shot based on the record of a camera", icon_previz.icon_id, 0),
-        ("STORYBOARD", "Storyboard Frame", "2D drawing used for storyboarding", icon_stb.icon_id, 1),
+        ("STORYBOARD", "Storyboard Shot", "2D drawing used for storyboarding", icon_stb.icon_id, 1),
     )
 
     return res
@@ -634,6 +634,11 @@ class UAS_ShotManager_Shot(ShotInterface, PropertyGroup):
             mode: "STORYBOARD"
         """
         # TODO: differenciate the modes of grease pencils to provide the right one
+        #     if self.isCameraValid():
+        # gp_child = utils_greasepencil.get_greasepencil_child(self.camera)
+        # if gp_child is not None:
+        #     gpProps = self.getGreasePencilProps(mode="STORYBOARD")
+
         gpProps = self.greasePencils[0] if len(self.greasePencils) else None
         return gpProps
 
@@ -712,6 +717,8 @@ class UAS_ShotManager_Shot(ShotInterface, PropertyGroup):
                 if props.use_greasepencil:
                     if "STORYBOARD" == self.shotType:
                         _showGreasePencil(gp_child, True)
+
+                    # PREVIZ
                     else:
                         if props.getCurrentShot() == self:
                             _showGreasePencil(gp_child, True)