+++ shotmanager/ui/sm_shots_ui_previz_layout.py
@@ -46,9 +46,6 @@ class UAS_UL_ShotManager_Items(bpy.types.UIList):
 
         display_getsetcurrentframe_in_shotlist = props.display_getsetcurrentframe_in_shotlist
 
-        currentIconIsOrange = True
-        orange = "_Orange" if currentIconIsOrange else ""
-        cam = f"Cam{orange}" if current_shot_index == index else ""
         currentFrame = context.scene.frame_current
 
         # check if the camera still exists in the scene
@@ -62,6 +59,15 @@ class UAS_UL_ShotManager_Items(bpy.types.UIList):
             numSharedCam = 2
 
         # draw the Duration components
+        ##########################
+
+        currentIconIsOrange = True
+        orange = "_Orange" if currentIconIsOrange else ""
+        if "PREVIZ" == item.shotType:
+            cam = f"Cam{orange}" if current_shot_index == index else ""
+        # STORYBOARD
+        else:
+            cam = f"Stb{orange}" if current_shot_index == index else "Stb"
 
         if item.enabled:
             icon = config.icons_col[f"ShotMan_Enabled{cam}"]